fix: Fixed custom fonts not loading

+++ App.js
@@ -1,6 +1,9 @@
 import React, { useState } from 'react';
 import { View, StatusBar, FlatList } from 'react-native';
 
+import { useFonts } from 'expo-font';
+import AppLoading from 'expo-app-loading';
+
 import styled from 'styled-components';
 import AddInput from './assets/components/AddInput';
 import Empty from './assets/components/Empty';
@@ -8,6 +11,11 @@ import Header from './assets/components/Header';
 import TodoList from './assets/components/TodoList';
 
 export default function App() {
+  const [fontsLoaded] = useFonts({
+    'poppins-bold': require('./assets/fonts/poppins-bold.ttf'),
+    'poppins-regular': require('./assets/fonts/poppins-regular.ttf'),
+  });
+
   const [data, setData] = useState([]);
   
   const submitHandler = (value) => {
@@ -28,27 +36,34 @@ export default function App() {
     });
   };
 
-  return (
-    <ComponentContainer>
-      <View>
-        <StatusBar barStyle='light-content' backgroundColor='midnightblue'></StatusBar>
-      </View>
-
-      <View>
-        <FlatList 
-          data={data} 
-          ListHeaderComponent={() => <Header></Header>} 
-          ListEmptyComponent={() => <Empty></Empty>}
-          keyExtractor={(item) => item.key} 
-          renderItem={({ item }) => (<TodoList item={item} 
-          deleteItem={deleteItem}></TodoList>)}>
-        </FlatList>
+  if (fontsLoaded) {
+    return (
+      <ComponentContainer>
         <View>
-          <AddInput submitHandler={submitHandler}></AddInput>
+          <StatusBar barStyle='light-content' backgroundColor='midnightblue'></StatusBar>
         </View>
-      </View>
-    </ComponentContainer>
-  );
+  
+        <View>
+          <FlatList 
+            data={data} 
+            ListHeaderComponent={() => <Header></Header>} 
+            ListEmptyComponent={() => <Empty></Empty>}
+            keyExtractor={(item) => item.key} 
+            renderItem={({ item }) => (<TodoList item={item} 
+            deleteItem={deleteItem}></TodoList>)}>
+          </FlatList>
+          <View>
+            <AddInput submitHandler={submitHandler}></AddInput>
+          </View>
+        </View>
+      </ComponentContainer>
+    );
+  }
+  else {
+    return (
+      <AppLoading></AppLoading>
+    );
+  }
 }
 
 const ComponentContainer = styled.View`
